% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TranscriptomHiCs.R
\name{add_annotation}
\alias{add_annotation}
\title{Add an annotation to a TranscriptomHiCs object.}
\usage{
add_annotation(x, annotation, annot_name)
}
\arguments{
\item{x}{A TranscriptomHiCs object.}
}
\description{
Add an annotation to a TranscriptomHiCs object.
}
