% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomHiCs.R
\name{annotate_count}
\alias{annotate_count}
\title{Adds an annotation count column to a GRanges object.}
\usage{
annotate_count(query_regions, annotation, label)
}
\description{
Adds an annotation count column to a GRanges object.
}
\keyword{internal}
