% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TranscriptomHiCs.R
\name{annotate_distant_binding}
\alias{annotate_distant_binding}
\title{Retrieves all binding types for the given annotation and annotates the promoters.}
\usage{
annotate_distant_binding(x, annot_name)
}
\arguments{
\item{x}{A TranscriptomHiCs object.}
}
\description{
Retrieves all binding types for the given annotation and annotates the promoters.
}
