% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomHiCs.R
\name{annotations}
\alias{annotations}
\title{Add an annotation to a GenomHiCs object.}
\usage{
annotations(x)
}
\arguments{
\item{x}{A GenomHiCs object.}
}
\description{
Add an annotation to a GenomHiCs object.
}
