% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomHiCs.R
\name{build_site_list}
\alias{build_site_list}
\title{Go from promoter/enhancer indices to a GRangesList of sites.}
\usage{
build_site_list(promoter_regions, promoter_ids, query_regions, query_ids)
}
\description{
Go from promoter/enhancer indices to a GRangesList of sites.
}
\keyword{internal}
