% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomHiCs.R
\docType{class}
\name{GenomHiCs-class}
\alias{GenomHiCs-class}
\title{GenomHiCs class.}
\description{
GenomHiCs class.
}
\section{Slots}{

\describe{
\item{\code{query_regions}}{query_regions.}

\item{\code{contacts}}{Contacts.}

\item{\code{annotations}}{Annotations}

\item{\code{structures}}{Structures}
}}

