% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TranscriptomHiCs.R
\name{get_all_sites}
\alias{get_all_sites}
\title{Get a list of all sites connected to the promoters.}
\usage{
get_all_sites(x, query_regions)
}
\arguments{
\item{x}{A TranscriptomHiCs object.}
}
\description{
Get a list of all sites connected to the promoters.
}
