% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TranscriptomHiCs.R
\name{get_direct_sites}
\alias{get_direct_sites}
\title{Get a list of site directly binding the promoters.}
\usage{
get_direct_sites(x, query_regions)
}
\arguments{
\item{x}{A TranscriptomHiCs object.}
}
\description{
Get a list of site directly binding the promoters.
}
