% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TranscriptomHiCs.R
\name{get_distant_sites}
\alias{get_distant_sites}
\title{Get a list of site distantly-bound to the promoters.}
\usage{
get_distant_sites(x, query_regions)
}
\arguments{
\item{x}{A TranscriptomHiCs object.}
}
\description{
Get a list of site distantly-bound to the promoters.
}
