% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TranscriptomHiCs.R
\name{get_window_sites}
\alias{get_window_sites}
\title{Get a list of site binding within the windows of the promoters.}
\usage{
get_window_sites(x, query_regions)
}
\arguments{
\item{x}{A TranscriptomHiCs object.}
}
\description{
Get a list of site binding within the windows of the promoters.
}
