% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TranscriptomHiCs.R
\name{is_structure_bound}
\alias{is_structure_bound}
\title{Determine if promoters are structure-bound by the given annotation.}
\usage{
is_structure_bound(x, colname)
}
\arguments{
\item{x}{A TranscriptomHiCs object.}
}
\description{
Determine if promoters are structure-bound by the given annotation.
}
