% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TranscriptomHiCs.R
\name{is_window_bound}
\alias{is_window_bound}
\title{Determine if promoters are window-bound by the given annotation.}
\usage{
is_window_bound(x, colname)
}
\arguments{
\item{x}{A TranscriptomHiCs object.}
}
\description{
Determine if promoters are window-bound by the given annotation.
}
