% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TranscriptomHiCs.R
\name{promoters}
\alias{promoters}
\title{Add an annotation to a TranscriptomHiCs object.}
\usage{
promoters(x)
}
\arguments{
\item{x}{A TranscriptomHiCs object.}
}
\description{
Add an annotation to a TranscriptomHiCs object.
}
