% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomHiCs.R
\name{query_regions}
\alias{query_regions}
\title{Add an annotation to a GenomHiCs object.}
\usage{
query_regions(x)
}
\arguments{
\item{x}{A GenomHiCs object.}
}
\description{
Add an annotation to a GenomHiCs object.
}
