% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TranscriptomHiCs.R
\docType{class}
\name{TranscriptomHiCs-class}
\alias{TranscriptomHiCs-class}
\title{TranscriptomHiCs class.}
\description{
TranscriptomHiCs class.
}
\section{Slots}{

\describe{
\item{\code{promoters}}{Promoters.}

\item{\code{contacts}}{Contacts.}

\item{\code{}}{}
}}

