% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TranscriptomHiCs.R
\name{transcriptomHiCs}
\alias{transcriptomHiCs}
\title{Build a TranscriptomHiCs object.}
\usage{
transcriptomHiCs(promoters, contacts, annotations, structures = NULL)
}
\description{
Build a TranscriptomHiCs object.
}
