% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomHiCs.R
\name{distant_binding}
\alias{distant_binding}
\title{Retrieves all binding types for the given annotation.}
\usage{
distant_binding(x, query_name, annot_name)
}
\arguments{
\item{x}{A GenomHiCs object.}
}
\description{
Retrieves all binding types for the given annotation.
}
