% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomHiCs.R
\name{get_distant_sites}
\alias{get_distant_sites}
\title{Get a list of site distantly-bound to the query_regions.}
\usage{
get_distant_sites(x, query_name, annot_name)
}
\arguments{
\item{x}{A GenomHiCs object.}
}
\description{
Get a list of site distantly-bound to the query_regions.
}
