% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomHiCs.R
\name{is_structure_bound}
\alias{is_structure_bound}
\title{Determine if query_regions are structure-bound by the given annotation.}
\usage{
is_structure_bound(x, query_name, annot_name)
}
\arguments{
\item{x}{A GenomHiCs object.}
}
\description{
Determine if query_regions are structure-bound by the given annotation.
}
