% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TranscriptomHiCs.R
\name{structures}
\alias{structures}
\title{Add an annotation to a TranscriptomHiCs object.}
\usage{
structures(x)
}
\arguments{
\item{x}{A TranscriptomHiCs object.}
}
\description{
Add an annotation to a TranscriptomHiCs object.
}
